refactor(starknet_infra_utils): move `cairo1_compiler_version` and its transitive closure
Few remarks:
- No code change, only move.
- I put the new file under "testing" feature since `Cargo.toml`-parsing
  code should not be ran in production.
- This is prep. for testing compiler version alignment between
  `Cargo.toml` and `build.rs`;
  should also be done for `cairo-native`compiler.
