+++ crates/blockifier_test_utils/Cargo.toml
@@ -10,21 +10,18 @@ description = "Test utilities for the blockifier."
 cairo_native = []
 
 [dependencies]
-cached.workspace = true
 cairo-lang-starknet-classes.workspace = true
 itertools.workspace = true
 pretty_assertions.workspace = true
 rstest.workspace = true
-serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true, features = ["arbitrary_precision"] }
 starknet-types-core.workspace = true
 starknet_api = { workspace = true, features = ["testing"] }
-starknet_infra_utils.workspace = true
+starknet_infra_utils = { workspace = true, features = ["testing"] }
 strum.workspace = true
 strum_macros.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
-toml.workspace = true
 tracing.workspace = true
 tracing-test.workspace = true
