+++ crates/starknet_infra_utils/Cargo.toml
@@ -7,22 +7,26 @@ license-file.workspace = true
 description = "Infrastructure utility."
 
 [features]
-testing = ["colored", "dep:assert-json-diff"]
+testing = ["cached", "colored", "dep:assert-json-diff", "toml"]
 
 [lints]
 workspace = true
 
 [dependencies]
 assert-json-diff = { workspace = true, optional = true }
+cached = { workspace = true, optional = true }
 colored = { workspace = true, optional = true }
-serde.workspace = true
+serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 tokio = { workspace = true, features = ["process", "rt", "time"] }
+toml = { workspace = true, optional = true }
 tracing.workspace = true
 
 [dev-dependencies]
+cached.workspace = true
 nix.workspace = true
 pretty_assertions.workspace = true
 rstest.workspace = true
 tokio = { workspace = true, features = ["macros", "rt", "signal", "sync"] }
+toml.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
